Refactor start up activity
