+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
     android:theme="@style/AppTheme.Launcher"
     tools:ignore="GoogleAppIndexingWarning">
     <activity
-      android:name=".activities.user.UserActivity">
+      android:name=".views.startup.StartUpView">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
         <category android:name="android.intent.category.LAUNCHER" />
