--- app/src/main/res/layout/activity_user.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  tools:context=".activities.user.UserActivity">
-
-  <RelativeLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    tools:layout_editor_absoluteX="150dp"
-    tools:layout_editor_absoluteY="16dp">
-
-    <android.support.design.widget.AppBarLayout
-      android:id="@+id/appBarLayout"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:background="@color/colorAccent"
-      android:fitsSystemWindows="true"
-      app:elevation="0dip"
-      app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
-
-      <android.support.v7.widget.Toolbar
-        android:id="@+id/toolbarAuth"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:titleTextColor="@color/colorPrimary" />
-    </android.support.design.widget.AppBarLayout>
-
-    <ScrollView
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:layout_below="@id/appBarLayout"
-      android:layout_marginTop="0dp"
-      android:fillViewport="true">
-
-      <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <LinearLayout
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:orientation="horizontal"
-          android:gravity="center">
-
-          <Button
-            android:id="@+id/signUp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:background="@color/colorAccent"
-            android:paddingTop="8dp"
-            android:paddingBottom="8dp"
-            android:text="@string/button_signUp"
-            android:textColor="@color/colorPrimary"
-            android:textSize="16sp" />
-
-          <Button
-            android:id="@+id/logIn"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            android:background="@color/colorAccent"
-            android:paddingTop="8dp"
-            android:paddingBottom="8dp"
-            android:text="@string/button_logIn"
-            android:textColor="@color/colorPrimary"
-            android:textSize="16sp" />
-
-        </LinearLayout>
-
-        <LinearLayout
-          android:id="@+id/fragment_container"
-          android:layout_width="match_parent"
-          android:layout_height="0dp"
-          android:layout_weight="1"
-          android:background="@color/colorPrimary"
-          android:orientation="vertical" />
-
-
-      </LinearLayout>
-    </ScrollView>
-
-  </RelativeLayout>
-
-</android.support.constraint.ConstraintLayout>
