+++ app/src/main/res/layout/login_fragment.xml
@@ -1,59 +1,57 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-  xmlns:android="http://schemas.android.com/apk/res/android"
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto"
+  xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  android:orientation="vertical"
-  android:padding="16dp"
-  android:gravity="center"
-  >
+  tools:context=".views.startup.StartUpView">
 
-  <TextView
-    android:id="@+id/login_view"
-    android:layout_width="wrap_content"
+
+  <android.support.design.widget.TextInputLayout
+    android:id="@+id/textInputLayout"
+    android:layout_width="0dp"
     android:layout_height="wrap_content"
-    android:text="@string/message_login"
-    android:textSize="18sp"
-    android:gravity="center"
-    />
-
+    android:layout_marginStart="8dp"
+    android:layout_marginTop="200dp"
+    android:layout_marginEnd="8dp"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toTopOf="parent">
 
     <android.support.design.widget.TextInputEditText
       android:id="@+id/fragment_userEmailLogin"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
-      android:layout_margin="8dp"
-      android:hint="@string/hint_userEmail"
-      android:inputType="text"
-      android:maxLength="25"
-      android:maxLines="1"
-      android:padding="8dp"
-      android:textColor="@color/colorPrimaryDark"
-      android:textSize="14sp"/>
+      android:hint="@string/hint_userEmail" />
+  </android.support.design.widget.TextInputLayout>
+
+  <android.support.design.widget.TextInputLayout
+    android:id="@+id/textInputLayout2"
+    android:layout_width="0dp"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="8dp"
+    android:layout_marginTop="16dp"
+    android:layout_marginEnd="8dp"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toBottomOf="@+id/textInputLayout">
 
     <android.support.design.widget.TextInputEditText
       android:id="@+id/fragment_userPasswordLogin"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
-      android:layout_margin="8dp"
-      android:hint="@string/hint_userPassword"
-      android:inputType="textPassword"
-      android:maxLength="25"
-      android:maxLines="1"
-      android:padding="8dp"
-      android:textColor="@color/colorPrimaryDark"
-      android:textSize="14sp"/>
+      android:hint="@string/hint_userPassword" />
+  </android.support.design.widget.TextInputLayout>
 
-    <Button
-      android:id="@+id/fragment_login"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_margin="16dp"
-      android:background="@color/colorAccent"
-      android:paddingTop="8dp"
-      android:paddingBottom="8dp"
-      android:text="@string/button_logIn"
-      android:textColor="@color/colorPrimary"
-      android:textSize="18sp" />
-
-</LinearLayout>
+  <Button
+    android:id="@+id/fragment_login"
+    android:layout_width="0dp"
+    android:layout_height="wrap_content"
+    android:layout_marginStart="32dp"
+    android:layout_marginTop="8dp"
+    android:layout_marginEnd="32dp"
+    android:text="@string/button_logIn"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintTop_toBottomOf="@+id/textInputLayout2" />
+</android.support.constraint.ConstraintLayout>
